\name{nestedRanksTest-package}
\alias{nestedRanksTest-package}
\docType{package}
\title{
Perform Mann-Whitney-Wilcoxon nonparametric comparison between two groups that are nested.
}
\description{
Description ...
}
\details{
\tabular{ll}{
Package: \tab nestedRanksTest\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-01-16\cr
License: \tab LGPL-3\cr
}
More details...
}
\seealso{\link{nestedRanksTest}}
\author{
Douglas G. Scofield

Maintainer: Douglas G. Scofield <douglasgscofield@gmail.com>
}
\references{
Thompson, P.G., P.E. Smouse, D.G. Scofield and V.L. Sork.  2014.  What seeds tell us about birds: a multi-year analysis of acorn woodpecker foraging movements.  \emph{Movement Ecology} 2, 12.  doi:10.1186/2051-3933-2-12, data at doi:10.5061/dryad.64jk6
}
\keyword{ package }
\keyword{ file }
\keyword{ manip }
\keyword{ attribute }
