% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nestedRanksTest.R
\name{nestedRanksTest_new}
\alias{nestedRanksTest_new}
\title{Potentially faster version of nestedRanksTest}
\usage{
nestedRanksTest_new(x, y, groups, n.iter = 10000, lightweight = FALSE, ...)
}
\description{
Potentially faster version of nestedRanksTest
}

