% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nestedRanksTest.R
\name{nestedRanksTest_Z}
\alias{nestedRanksTest_Z}
\title{Calculates Z-score from ranks.}
\usage{
nestedRanksTest_Z(x, n1, n2)
}
\arguments{
\item{x}{Values to be ranked for the test.  Its length must
be equal to the sum of \code{n1} and \code{n2}.}

\item{n1}{The first \code{n1} values in \code{x} belong to the
first treatment level.}

\item{n2}{The final \code{n2} values in \code{x} belong to the
            second treatment level.}
}
\value{
The calculated Z-score
}
\description{
\code{nestedRanksTest_Z} is used by \code{nestedRanksTest} to
calculate the Z-score for the ranks of \code{x} divided into two
treatment levels.
}
\details{
Values across both treatments are ranked using the base R function
\code{rank} with \code{ties.method = "average"}, which assigns
tied values their average rank.  The Mann-Whitney-Wilcoxon test
statistic is then computed.  Because the scale of this statistic
is sample-size dependent (between \eqn{- n_1 n_2} and
\eqn{+ n_1 n_2}), it is scaled to be \eqn{[-1, +1]} by dividing by
\eqn{n_1 n_2}.
}
\seealso{
\code{\link{nestedRanksTest}}, \code{\link{wilcox.test}}
}
\keyword{internal}

