% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nestedRanksTest.R
\name{nestedRanksTest_Z}
\alias{nestedRanksTest_Z}
\title{Calculates Z-score from ranks.}
\usage{
nestedRanksTest_Z(y, n1, n2)
}
\arguments{
\item{y}{Values to be ranked for the test.  Its length must
be equal to the sum of \code{n1} and \code{n2}.}

\item{n1}{The first \code{n1} values in \code{y} belong to the
first treatment level.}

\item{n2}{The final \code{n2} values in \code{y} belong to the
            second treatment level.}
}
\value{
The calculated Z-score
}
\description{
\code{nestedRanksTest_Z} is used by \code{nestedRanksTest} to
calculate the Z-score for the ranks of responses \code{y} divided
into two treatment levels.
}
\details{
Values across both treatments are ranked using the base R function
\code{rank} with \code{ties.method = "average"}, which assigns
tied values their average rank.  The Mann-Whitney-Wilcoxon test
statistic is computed from these ranks.  Because the value of the
statistic is sample-size dependent (between \code{-n1*n2} and
\code{n1*n2}), it is scaled to be \code{[-1, +1]} by dividing by
\code{n1*n2}.

The bottleneck here is calculation of ranks, so speedup of this
function comes from speeding up ranks().  A stripped-down rank_new:

function (x) y <- .Internal(rank(x, length(x), "average"))

is much faster than the default rank:

> microbenchmark(rank(yy), rank_new(yy), times=100000)
    Unit: microseconds
             expr    min     lq      mean median      uq       max neval
         rank(yy) 29.148 31.945 36.931556 32.678 33.5165 57192.350 1e+05
     rank_new(yy)  3.755  4.300  4.789952  4.542  4.7290  6784.741 1e+05
}
\seealso{
\code{\link{nestedRanksTest}}, \code{\link{wilcox.test}}
}

